+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unrleased]
+
+### Fixed
+
+- `wal-scratch-persist` no longer persists the
+  `initial-scratch-message`.
+
 ## [v2.2.0]
 
 Prelude to peace.